requirements.txt
pymongo==4.6.0
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
watchdog==3.0.0
dash==2.14.2
dash-bootstrap-components==1.5.0
